+++ Editor/Exporter/ExportSettingsEditor.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.IO;
-using Treasured.UnitySdk;
-using UnityEditor;
+﻿using UnityEditor;
 using UnityEngine;
 
 namespace Treasured.UnitySdk
@@ -12,25 +9,14 @@ namespace Treasured.UnitySdk
         public override void OnInspectorGUI()
         {
             ExportSettings settings = target as ExportSettings;
-            EditorGUI.BeginChangeCheck();
-            string newOutputFolderName = EditorGUILayout.TextField(new GUIContent("Output Folder Name"), settings.folderName);
-            if (EditorGUI.EndChangeCheck() && !string.IsNullOrWhiteSpace(newOutputFolderName))
+            EditorGUILayout.LabelField("Export Settings", EditorStyles.boldLabel);
+            using(new EditorGUI.IndentLevelScope(1))
             {
-                settings.folderName = newOutputFolderName;
-            }
-            using (new EditorGUILayout.HorizontalScope())
-            {
-                EditorGUILayout.PrefixLabel(GUIContent.none);
-                if (GUILayout.Button(new GUIContent("Open root folder", "Open the root output folder in the File Explorer. The default output root will be the path of your Unity project/Treasured Data"), GUILayout.Height(24)))
+                EditorGUI.BeginChangeCheck();
+                string newOutputFolderName = EditorGUILayout.TextField(new GUIContent("Folder Name"), settings.folderName);
+                if (EditorGUI.EndChangeCheck() && !string.IsNullOrWhiteSpace(newOutputFolderName))
                 {
-                    Application.OpenURL(ExportSettings.DefaultOutputDirectory);
-                }
-                using (new EditorGUI.DisabledGroupScope(!Directory.Exists(settings.OutputDirectory)))
-                {
-                    if (GUILayout.Button(new GUIContent("Open current folder", "Open the current output folder in the File Explorer. This option is available after first export."), GUILayout.Height(24)))
-                    {
-                        Application.OpenURL(settings.OutputDirectory);
-                    }
+                    settings.folderName = newOutputFolderName;
                 }
             }
         }